@@ -31,6 +31,7 @@ namespace FunFair.CodeAnalysis.Tests.Verifiers
         ///     Returns the codefix being tested (VB) - to be implemented in non-abstract class
         /// </summary>
         /// <returns>The CodeFixProvider to be used for VisualBasic code</returns>
+        // ReSharper disable once UnusedMember.Global
         protected virtual CodeFixProvider? GetBasicCodeFixProvider()
         {
             return null;
@@ -43,6 +44,7 @@ namespace FunFair.CodeAnalysis.Tests.Verifiers
         /// <param name="newSource">A class in the form of a string after the CodeFix was applied to it</param>
         /// <param name="codeFixIndex">Index determining which codefix to apply if there are multiple</param>
         /// <param name="allowNewCompilerDiagnostics">A bool controlling whether or not the test will fail if the CodeFix introduces other warnings after being applied</param>
+        // ReSharper disable once UnusedMember.Global
         protected Task VerifyCSharpFixAsync(string oldSource, string newSource, int? codeFixIndex = null, bool allowNewCompilerDiagnostics = false)
         {
             DiagnosticAnalyzer? analyzer = this.GetCSharpDiagnosticAnalyzer();
@@ -96,11 +98,8 @@ namespace FunFair.CodeAnalysis.Tests.Verifiers
 
             for (int i = 0; i < attempts; ++i)
             {
-                List<CodeAction> actions = new List<CodeAction>();
-                CodeFixContext context = new CodeFixContext(document: document,
-                                                            analyzerDiagnostics[0],
-                                                            registerCodeFix: (a, d) => actions.Add(a),
-                                                            cancellationToken: CancellationToken.None);
+                List<CodeAction> actions = new();
+                CodeFixContext context = new(document: document, analyzerDiagnostics[0], registerCodeFix: (a, _) => actions.Add(a), cancellationToken: CancellationToken.None);
                 await codeFixProvider.RegisterCodeFixesAsync(context);
 
                 if (!actions.Any())
