@@ -20,12 +20,12 @@ namespace FunFair.CodeAnalysis
 
         private static readonly ProhibitedMethodsSpec[] ForcedMethods =
         {
-            new ProhibitedMethodsSpec(ruleId: Rules.RuleDontUseSubstituteReceivedWithZeroNumberOfCalls,
-                                      title: "Avoid use of received with zero call count",
-                                      message: "Only use Received with expected call count greater than 0, use DidNotReceived instead if 0 call received expected",
-                                      sourceClass: "NSubstitute.SubstituteExtensions",
-                                      forcedMethod: "Received",
-                                      new[] {new[] {new ParameterSpec(name: "requiredNumberOfCalls", type: "NumericLiteralExpression", value: "0")}})
+            new(ruleId: Rules.RuleDontUseSubstituteReceivedWithZeroNumberOfCalls,
+                title: "Avoid use of received with zero call count",
+                message: "Only use Received with expected call count greater than 0, use DidNotReceived instead if 0 call received expected",
+                sourceClass: "NSubstitute.SubstituteExtensions",
+                forcedMethod: "Received",
+                new[] {new[] {new ParameterSpec(name: "requiredNumberOfCalls", type: "NumericLiteralExpression", value: "0")}})
         };
 
         /// <inheritdoc />
@@ -64,7 +64,7 @@ namespace FunFair.CodeAnalysis
                         continue;
                     }
 
-                    Mapping mapping = new Mapping(methodName: memberSymbol.Name, SymbolDisplay.ToDisplayString(memberSymbol.ContainingType));
+                    Mapping mapping = new(methodName: memberSymbol.Name, SymbolDisplay.ToDisplayString(memberSymbol.ContainingType));
 
                     IEnumerable<ProhibitedMethodsSpec> forcedMethods = ForcedMethods.Where(predicate: rule => rule.QualifiedName == mapping.QualifiedName);
 
@@ -142,12 +142,7 @@ namespace FunFair.CodeAnalysis
 
         private sealed class ProhibitedMethodsSpec
         {
-            public ProhibitedMethodsSpec(string ruleId,
-                                         string title,
-                                         string message,
-                                         string sourceClass,
-                                         string forcedMethod,
-                                         IEnumerable<IEnumerable<ParameterSpec>> bannedSignatures)
+            public ProhibitedMethodsSpec(string ruleId, string title, string message, string sourceClass, string forcedMethod, IEnumerable<IEnumerable<ParameterSpec>> bannedSignatures)
             {
                 this.SourceClass = sourceClass;
                 this.ForcedMethod = forcedMethod;
