@@ -43,7 +43,7 @@ public sealed class Test {
     }
 }";
 
-            DiagnosticResult expected = new DiagnosticResult
+            DiagnosticResult expected = new()
                                         {
                                             Id = "FFS0024",
                                             Message = "ILogger parameters on leaf classes should not be ILogger but ILogger<Test>",
@@ -67,7 +67,7 @@ public abstract class Test {
     }
 }";
 
-            DiagnosticResult expected = new DiagnosticResult
+            DiagnosticResult expected = new()
                                         {
                                             Id = "FFS0023",
                                             Message = "ILogger parameters on base classes should not be ILogger<Test> but ILogger",
@@ -111,7 +111,7 @@ public sealed class Test {
     }
 }";
 
-            DiagnosticResult expected = new DiagnosticResult
+            DiagnosticResult expected = new()
                                         {
                                             Id = "FFS0025",
                                             Message = "Should be using 'Test' rather than 'Banana' with Microsoft.Extensions.Logging.ILogger<TCategoryName>",
