@@ -41,7 +41,7 @@ using Xunit;
             {
             }
 }";
-            DiagnosticResult expected = new DiagnosticResult
+            DiagnosticResult expected = new()
                                         {
                                             Id = "FFS0013",
                                             Message = "Test classes should be derived from TestBase",
@@ -74,12 +74,7 @@ using Xunit.Abstractions;
 }";
 
             return this.VerifyCSharpDiagnosticAsync(source: test,
-                                                    new[]
-                                                    {
-                                                        WellKnownMetadataReferences.Xunit,
-                                                        WellKnownMetadataReferences.FunFairTestCommon,
-                                                        WellKnownMetadataReferences.XunitAbstractions
-                                                    });
+                                                    new[] {WellKnownMetadataReferences.Xunit, WellKnownMetadataReferences.FunFairTestCommon, WellKnownMetadataReferences.XunitAbstractions});
         }
 
         [Fact]
@@ -114,7 +109,7 @@ using Xunit;
             {
             }
 }";
-            DiagnosticResult expected = new DiagnosticResult
+            DiagnosticResult expected = new()
                                         {
                                             Id = "FFS0013",
                                             Message = "Test classes should be derived from TestBase",
@@ -148,12 +143,7 @@ using Xunit.Abstractions;
 }";
 
             return this.VerifyCSharpDiagnosticAsync(source: test,
-                                                    new[]
-                                                    {
-                                                        WellKnownMetadataReferences.Xunit,
-                                                        WellKnownMetadataReferences.FunFairTestCommon,
-                                                        WellKnownMetadataReferences.XunitAbstractions
-                                                    });
+                                                    new[] {WellKnownMetadataReferences.Xunit, WellKnownMetadataReferences.FunFairTestCommon, WellKnownMetadataReferences.XunitAbstractions});
         }
 
         [Fact]
