@@ -18,20 +18,20 @@ namespace FunFair.CodeAnalysis
 
         private static readonly TypeCheckSpec[] Specifications =
         {
-            new TypeCheckSpec(ruleId: Rules.LoggerParametersOnBaseClassesShouldNotUseGenericLoggerCategory,
-                              title: "ILogger parameters on base classes should not be ILogger<{0}> but ILogger",
-                              message: "ILogger parameters on base classes should not be ILogger<{0}> but ILogger",
-                              allowedSourceClass: "Microsoft.Extensions.Logging.ILogger",
-                              prohibitedClass: "Microsoft.Extensions.Logging.ILogger<TCategoryName>",
-                              isProtected: true,
-                              matchTypeOnGenericParameters: false),
-            new TypeCheckSpec(ruleId: Rules.LoggerParametersOnLeafClassesShouldUseGenericLoggerCategory,
-                              title: "ILogger parameters on leaf classes should not be ILogger but ILogger<{0}>",
-                              message: "ILogger parameters on leaf classes should not be ILogger but ILogger<{0}>",
-                              allowedSourceClass: "Microsoft.Extensions.Logging.ILogger<TCategoryName>",
-                              prohibitedClass: "Microsoft.Extensions.Logging.ILogger",
-                              isProtected: false,
-                              matchTypeOnGenericParameters: true)
+            new(ruleId: Rules.LoggerParametersOnBaseClassesShouldNotUseGenericLoggerCategory,
+                title: "ILogger parameters on base classes should not be ILogger<{0}> but ILogger",
+                message: "ILogger parameters on base classes should not be ILogger<{0}> but ILogger",
+                allowedSourceClass: "Microsoft.Extensions.Logging.ILogger",
+                prohibitedClass: "Microsoft.Extensions.Logging.ILogger<TCategoryName>",
+                isProtected: true,
+                matchTypeOnGenericParameters: false),
+            new(ruleId: Rules.LoggerParametersOnLeafClassesShouldUseGenericLoggerCategory,
+                title: "ILogger parameters on leaf classes should not be ILogger but ILogger<{0}>",
+                message: "ILogger parameters on leaf classes should not be ILogger but ILogger<{0}>",
+                allowedSourceClass: "Microsoft.Extensions.Logging.ILogger<TCategoryName>",
+                prohibitedClass: "Microsoft.Extensions.Logging.ILogger",
+                isProtected: false,
+                matchTypeOnGenericParameters: true)
         };
 
         private static readonly DiagnosticDescriptor MissMatchTypes = RuleHelpers.CreateRule(code: Rules.GenericTypeMissMatch,
@@ -93,8 +93,7 @@ namespace FunFair.CodeAnalysis
                 return;
             }
 
-            TypeCheckSpec? rule =
-                Specifications.FirstOrDefault(ns => ns.IsProtected == isProtected && (ns.AllowedSourceClass == fullTypeName || ns.ProhibitedClass == fullTypeName));
+            TypeCheckSpec? rule = Specifications.FirstOrDefault(ns => ns.IsProtected == isProtected && (ns.AllowedSourceClass == fullTypeName || ns.ProhibitedClass == fullTypeName));
 
             if (rule == null)
             {
@@ -105,10 +104,7 @@ namespace FunFair.CodeAnalysis
             {
                 if (rule.MatchTypeOnGenericParameters)
                 {
-                    CheckGenericParameterTypeMatch(syntaxNodeAnalysisContext: syntaxNodeAnalysisContext,
-                                                   parameterSyntax: parameterSyntax,
-                                                   className: className,
-                                                   fullTypeName: fullTypeName);
+                    CheckGenericParameterTypeMatch(syntaxNodeAnalysisContext: syntaxNodeAnalysisContext, parameterSyntax: parameterSyntax, className: className, fullTypeName: fullTypeName);
                 }
 
                 return;
@@ -120,10 +116,7 @@ namespace FunFair.CodeAnalysis
             }
         }
 
-        private static void CheckGenericParameterTypeMatch(SyntaxNodeAnalysisContext syntaxNodeAnalysisContext,
-                                                           ParameterSyntax parameterSyntax,
-                                                           string className,
-                                                           string fullTypeName)
+        private static void CheckGenericParameterTypeMatch(SyntaxNodeAnalysisContext syntaxNodeAnalysisContext, ParameterSyntax parameterSyntax, string className, string fullTypeName)
         {
             IParameterSymbol? ds = syntaxNodeAnalysisContext.SemanticModel.GetDeclaredSymbol(parameterSyntax);
 
@@ -142,11 +135,7 @@ namespace FunFair.CodeAnalysis
 
                         if (displayName != className)
                         {
-                            syntaxNodeAnalysisContext.ReportDiagnostic(Diagnostic.Create(descriptor: MissMatchTypes,
-                                                                                         parameterSyntax.GetLocation(),
-                                                                                         className,
-                                                                                         displayName,
-                                                                                         fullTypeName));
+                            syntaxNodeAnalysisContext.ReportDiagnostic(Diagnostic.Create(descriptor: MissMatchTypes, parameterSyntax.GetLocation(), className, displayName, fullTypeName));
                         }
                     }
                 }
@@ -155,13 +144,7 @@ namespace FunFair.CodeAnalysis
 
         private sealed class TypeCheckSpec
         {
-            public TypeCheckSpec(string ruleId,
-                                 string title,
-                                 string message,
-                                 string allowedSourceClass,
-                                 string prohibitedClass,
-                                 bool isProtected,
-                                 bool matchTypeOnGenericParameters)
+            public TypeCheckSpec(string ruleId, string title, string message, string allowedSourceClass, string prohibitedClass, bool isProtected, bool matchTypeOnGenericParameters)
             {
                 this.AllowedSourceClass = allowedSourceClass;
                 this.ProhibitedClass = prohibitedClass;
